containers/crawler/requirements.txt
# Scraping
requests
html5lib
bs4
# building
build
# Testing
coverage
hypothesis
pytest
# Database
pymongo
# ENV 
python-dotenv